Initial Commit. Total hours: 12.5
Discussing which project to work on with Chris (site lead)
Getting RTC set up on machine
Discussing when to work on project and who to talk to
Figuring out log-in information and downloading necessary programs.
In-depth discussion with Ryan (resident RTC expert) on the inner workings of RTC and the ALM suite. It streamlines Git, JIRA, DOORS, Dimensions, and other things!
Investigation into how RTC works and how to pull from our version of it. Figured out basic commands for pulling from a stream.